Refactored into multiple classes
